Refactor parsing IEvent's from XML
Fixes a bug where the XML parsing started reading junk data, which
occasionally caused some settings (start time type and duration type)
to get reset upon opening a DISSCO project.

The XML parsing is still extremely brittle (and should be replaced in
the future), but is now just a little bit more robust than before.

+++ LASSIE/src/ProjectViewController.cpp
@@ -1807,14 +1807,14 @@ ProjectViewController::ProjectViewController(MainWindow* _mainWindow, std::strin
         std::stringstream ss(data);
 
         // read the number of models
-        int size;
+        size_t size;
         ss >> size;
         markovModels.resize(size);
 
         // read individual models
         string modelText, line;
         getline(ss, line, '\n');
-        for (int i = 0; i < size; i++) {
+        for (size_t i = 0; i < size; i++) {
             getline(ss, line, '\n');
             modelText = line + '\n';
             getline(ss, line, '\n');
